Add SimpleH264Encoder by Jordi Cenzano
Ported to C# by Roger Hardiman with some help from the free trial version
of Tangible Software's C++ to C# conversion tool.

This H264 Encoder will generate the SPS and PPS but DOES NOT COMPRESS
the video. It takes raw YUV data and wraps it in a lossless H264 encoding.
The resulting data is slightly larger than the original YUV data.

A real project would use a proper H264 encoder that compresses the data.
SimpleH264Encoder is used to get a quick demonstration in 100% C# of
a H264 compliant bit stream.