Panasonic HE40 does not include SPS/PPS in the SDP
Allow the output .264 file to be opened when NALs are received (which will have in-band SPS and PPS
data).

All the other cameras tested include the H264 SPS and PPS out-of-band in the DESCRIBE
message SDP data.

Users of the library need to be aware that the SPS/PPS Event will not be Fired on the Panasonic camera
and will have to extract the SPS and PPS from the NALs