Add UDP streaming to RTSP Camera Example
Will send video via UDP.
If there is a TEARDOWN message it will stop sending.

TODO - check if Client/Viewer is still connected by checking when the last
RTSP message was received and when the last RTCP message was received.
If they timeout, then stop sending. That is for a future commit